+++ bit/tests/test_gen_kraken2_tax_plots.py
@@ -1,4 +1,3 @@
-import io
 from pathlib import Path
 import pandas as pd
 import pytest